# configs/swagger/v1.yaml
swagger: "2.0"
info:
  description: "悟空聊天IM API"
  version: "1.0.0"
  title: "悟空聊天IM API"
host: "49.235.59.182:8080"
...
schemes:
  - "http"
basePath: "/v1"
paths:
...
  /label:
    post:
      tags:
        - "label"
      summary: "添加标签"
      description: "添加标签"
      operationId: "labelAdd"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: 'body'
          description: "标签内容"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: "标签名称"
              member_uids:
                 type: array
                 items:
                   type: string
                 description: "成员uid集合" 
      responses:
        200:
          description: "成功"
          schema:
           $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
    get:
      tags:
        - "label"
      summary: "标签列表"
      description: "标签列表"
      operationId: "labelList"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "成功"
          schema:
            type: array
            items:
              properties:
                id:
                 type: integer
                 description: '标签ID'
                name:
                 type: string
                 description: '标签名称'
                count:
                 type: integer
                 description: '成员数量'
                members:
                  type: array
                  items:
                    properties:
                     uid:
                      type: string
                      description: "成员ID"
                     name:
                      type: string
                      description: "成员名称"
                     remark:
                      type: string
                      description: '成员备注'
                
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: [] 
  /label/{id}:
    put:
      tags:
        - "label"
      summary: "修改标签"
      description: "修改标签"
      operationId: "labelUpdate"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "标签ID"
          required: true
          type: "integer"
        - in: "body"
          name: 'body'
          description: "标签内容"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: "标签名称"
              member_uids:
                 type: array
                 items:
                   type: string
                 description: "成员uid集合" 
      responses:
        200:
          description: "成功"
          schema:
           $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
    delete:
      tags:
        - "label"
      summary: "删除标签"
      description: "删除标签"
      operationId: "labelDelete"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "标签ID"
          required: true
          type: "integer"
      responses:
        200:
          description: "成功"
          schema:
           $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
    get:
      tags:
        - "label"
      summary: "标签详情"
      description: "删除详情"
      operationId: "labelDetial"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "标签ID"
          required: true
          type: "integer"
      responses:
        200:
          description: "成功"
          schema:
            type: object
            properties:
             id:
              type: integer
              description: '标签ID'
             name:
              type: string
              description: '标签名称'
             count:
              type: integer
              description: '标签成员数量'
             members:
              type: array
              items:
                 properties:
                  uid:
                   type: string
                   description: "成员ID"
                  name:
                   type: string
                   description: "成员名称"
                  remark:
                   type: string
                   description: '成员备注'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []    
securityDefinitions:
  token:
    type: "apiKey"
    in: "header"
    name: "token"
    description: "用户token"
definitions:
...
  response:
    type: "object"
    properties:
      status:
        type: integer
        format: int
      msg:
        type: "string"